clean up panel configs
